Fixed issue with "Max rows" filter always being "query optimized" even in output data streams where there is no real query going on.
